refactor(config): generate SessionInitConfig from session config macro
Replace the hand-written SessionInitConfig, whose fields duplicated the
streaming parallelism family of SessionConfig, with a struct generated by
the SessionConfig derive macro. Parameters opt in via a new SESSION_INIT
flag, so adding a future init-able parameter is a one-line flag change and
the field names, types, and docs stay in sync with SessionConfig.
