+++ src/common/src/config/mod.rs
@@ -32,8 +32,8 @@ pub mod streaming;
 pub use streaming::{AsyncStackTraceOption, StreamingConfig};
 pub mod server;
 pub use server::{HeapProfilingConfig, ServerConfig};
-pub mod session_init;
-pub use session_init::SessionInitConfig;
+
+pub use crate::session_config::SessionInitConfig;
 pub mod udf;
 pub use udf::UdfConfig;
 pub mod storage;
@@ -507,6 +507,46 @@ pub mod tests {
         let expected = rw_config_to_markdown();
         let actual = expect_test::expect_file!["../../../config/docs.md"];
         actual.assert_eq(&expected);
+    }
+
+    #[test]
+    fn test_session_init_entries_distinguishes_omitted_from_default() {
+        let config: RwConfig = toml::from_str(
+            r#"
+            [session_init]
+            streaming_parallelism = "bounded(8)"
+            streaming_parallelism_for_table = "default"
+            "#,
+        )
+        .unwrap();
+
+        // Omitted fields are `None`; an explicit `default` is `Some("default")`.
+        assert_eq!(
+            config.session_init.streaming_parallelism.as_deref(),
+            Some("bounded(8)")
+        );
+        assert_eq!(
+            config
+                .session_init
+                .streaming_parallelism_for_table
+                .as_deref(),
+            Some("default")
+        );
+        assert_eq!(config.session_init.streaming_parallelism_for_sink, None);
+
+        // Only explicitly-configured parameters are reported, by their session parameter name.
+        assert_eq!(
+            config.session_init.entries(),
+            vec![
+                ("streaming_parallelism", "bounded(8)"),
+                ("streaming_parallelism_for_table", "default"),
+            ]
+        );
+    }
+
+    #[test]
+    fn test_session_init_default_is_empty() {
+        assert!(SessionInitConfig::default().entries().is_empty());
     }
 
     #[derive(Debug)]
