+++ src/config/docs.md
@@ -200,13 +200,13 @@ This page is automatically generated by `./risedev generate-example-config`
 
 | Config | Description | Default |
 |--------|-------------|---------|
-| streaming_parallelism |  | "" |
-| streaming_parallelism_for_backfill |  | "" |
-| streaming_parallelism_for_index |  | "" |
-| streaming_parallelism_for_materialized_view |  | "" |
-| streaming_parallelism_for_sink |  | "" |
-| streaming_parallelism_for_source |  | "" |
-| streaming_parallelism_for_table |  | "" |
+| streaming_parallelism | The execution parallelism for streaming queries, including tables, materialized views, indexes, and sinks. Defaults to `default`, which preserves the legacy adaptive scheduling behavior during effective resolution. | "" |
+| streaming_parallelism_for_backfill | Specific parallelism for backfill. Only `default` and a fixed positive integer are supported here. Adaptive backfill strategies are deferred to a later change. | "" |
+| streaming_parallelism_for_index | Specific parallelism for index. By default, it will fall back to `STREAMING_PARALLELISM`. | "" |
+| streaming_parallelism_for_materialized_view | Specific parallelism for materialized view. By default, it will fall back to `STREAMING_PARALLELISM`. | "" |
+| streaming_parallelism_for_sink | Specific parallelism for sink. By default, it will fall back to `STREAMING_PARALLELISM`. | "" |
+| streaming_parallelism_for_source | Specific parallelism for source. Defaults to `default`, which preserves the legacy bounded adaptive behavior only when the global parallelism itself remains `default`. Otherwise it follows the explicit global parallelism. | "" |
+| streaming_parallelism_for_table | Specific parallelism for table. Defaults to `default`, which preserves the legacy bounded adaptive behavior only when the global parallelism itself remains `default`. Otherwise it follows the explicit global parallelism. | "" |
 
 ## storage
