+++ src/config/example.toml
@@ -488,7 +488,7 @@ enable_embedded_wasm_udf = true
 streaming_parallelism = ""
 streaming_parallelism_for_backfill = ""
 streaming_parallelism_for_table = ""
-streaming_parallelism_for_source = ""
 streaming_parallelism_for_sink = ""
 streaming_parallelism_for_index = ""
+streaming_parallelism_for_source = ""
 streaming_parallelism_for_materialized_view = ""
